Changed the theme of the game
I had initially created the name of the game with no real direction of what is going on. I have a vision for the game, and the name "Nightfall Skirmish" never really fit it. Instead, this will be the new face of the game that I am currently working on.