Finalized level 1
I intend to leave this project as is, being a smaller yet completed game. The goal was to make a simple 3rd person shooter with some basic enemy interaction and some other functionality. You can compete with yourself to beat your record time for the level.

The code has been written in a way where if I decide to I will implement levels further down the line. Thanks for checking out my project in the mean time!